+++ src/AlgoCraft/DesgasteDeFuerza.java
@@ -2,7 +2,7 @@ package AlgoCraft;
 
 public class DesgasteDeFuerza extends Desgaste{
     @Override
-    public int desgastarDada(int durabilidad, int fuerza) {
+    public double desgastarDada(double durabilidad, int fuerza)  {
         return (durabilidad - fuerza);
     }
 }
